requirements.txt
Flask==2.1.2
pdfplumber==0.7.4
PyPDF2==1.26.0
python-docx==0.8.11
fuzzywuzzy==0.18.0
Werkzeug==2.1.2